requirements.txt
fastapi
uvicorn
pydantic
sentence-transformers
faiss-cpu
python-multipart
pypdf
slowapi
openai
